+++ CHANGELOG.md
@@ -32,6 +32,7 @@ Line wrap the file at 100 chars.                                              Th
 - WireGuard key page now shows a label explaining why buttons are disabled when in a blocked state
 - WireGuard key generation will try to replace old key if one exists.
 - Show banner about new app versions only if current platform has changes in latest release.
+- Don't make a GeoIP lookup by default in CLI status command. Add --location flag for enabling it.
 
 ### Fixed
 - Fix old settings deserialization to allow migrating settings from versions older than 2019.6.